posixpath again
